manus_local_no_n8n/requirements.txt
ollama
python-dotenv
playwright
pyautogui
pygetwindow
pyperclip
psutil
mss
pillow
fastapi
uvicorn
